Crate debris_backends

Backend implementations for the debris compiler

A Backend trait has to handle LLIR and produce a Directory.

The module common defines useful items that are specifc to minecraft.

Modules

common
Defines common items that are specific to the latest minecraft version

Structs

DatapackBackend
The Datapack Backend implementation

Traits

Backend
A Backend for debris, which has to convert LLIR into a Directory