Enum debris_backends::common::commands::MinecraftCommand

pub enum MinecraftCommand {
    ScoreboardSet {
        player: ScoreboardPlayer,
        value: i32,
    },
    ScoreboardSetEqual {
        player1: ScoreboardPlayer,
        player2: ScoreboardPlayer,
    },
    ScoreboardSetFromResult {
        player: ScoreboardPlayer,
        command: Box<MinecraftCommand>,
    },
    ScoreboardOperation {
        player1: ScoreboardPlayer,
        player2: ScoreboardPlayer,
        operation: ScoreboardOperation,
    },
    ScoreboardOperationAdd {
        player: ScoreboardPlayer,
        value: i32,
    },
    Execute {
        parts: Vec<ExecuteComponent>,
        and_then: Option<Box<MinecraftCommand>>,
    },
    Function {
        function: Rc<FunctionIdent>,
    },
    ScoreboardAdd {
        name: Rc<str>,
        criterion: ObjectiveCriterion,
        json_name: Option<String>,
    },
    ScoreboardRemove {
        name: Rc<str>,
    },
    RawCommand {
        command: Rc<str>,
    },
    JsonMessage {
        target: WriteTarget,
        message: String,
    },
}

Enumerates all minecraft commands that are used by any backend

Variants

ScoreboardSet

Fields

player: ScoreboardPlayer

value: i32

Sets the scoreboard value to a specific integer

ScoreboardSetEqual

Fields

player1: ScoreboardPlayer

player2: ScoreboardPlayer

Sets the scoreboard value to another scoreboard value

ScoreboardSetFromResult

Fields

player: ScoreboardPlayer

command: Box<MinecraftCommand>

Sets the scoreboard value equal to the result of the other command

ScoreboardOperation

Fields

player1: ScoreboardPlayer

player2: ScoreboardPlayer

operation: ScoreboardOperation

ScoreboardOperationAdd

Fields

player: ScoreboardPlayer

value: i32

Quick operation which adds or removes a static value

Execute

Fields

parts: Vec<ExecuteComponent>

and_then: Option<Box<MinecraftCommand>>

Any used execute command

Function

Fields

function: Rc<FunctionIdent>

Calls another function

ScoreboardAdd

Fields

name: Rc<str>

criterion: ObjectiveCriterion

json_name: Option<String>

ScoreboardRemove

Fields

name: Rc<str>

RawCommand

Fields

command: Rc<str>

JsonMessage

Fields

target: WriteTarget

message: String