Module debris_backends::datapack

A Backend that can compile to minecraft datapacks

Re-exports

pub use backend::DatapackBackend;

Modules

backend

function_context

generator

json_formatter

scoreboard_constants

scoreboard_context

stringify

templates

Implements some functions which can static files which never really change except for some parameters, like ‘pack.mcmeta’, ‘tick.json’ and so on

Structs

Datapack

Represents an in-memory datapack