Struct debris_backends::datapack::generator::DatapackGenerator

pub struct DatapackGenerator<'a> {
    compile_context: &'a CompileContext,
    llir: &'a Llir,
    function_calls_stats: FxHashMap<BlockId, usize>,
    function_ctx: FunctionContext,
    stack: Vec<Vec<MinecraftCommand>>,
    scoreboard_ctx: ScoreboardContext,
    scoreboard_constants: ScoreboardConstants,
}

This struct is used to generate a datapack from the llir representation

Fields
compile_context: &'a CompileContext

The compilation configuration

llir: &'a Llir

The llir to compile

function_calls_stats: FxHashMap<BlockId, usize>

Statistics for how often each function got called

function_ctx: FunctionContext

Contains the already generated functions

stack: Vec<Vec<MinecraftCommand>>

The current stack

Commands are pushed into the last value of last context

scoreboard_ctx: ScoreboardContext

A context which keeps track of the currently used scoreboards

scoreboard_constants: ScoreboardConstants

Keeps track of all constants that are used throughout the code

Implementations

impl<'a> DatapackGenerator<'a>

fn add_command(&mut self, command: MinecraftCommand)

Adds a command to the current stack

fn catch_ouput(&mut self, node: &Node) -> Vec<MinecraftCommand>

Handles the given command and returns the produced output

fn handle_main_function(
&mut self,
block_ids: impl Iterator<Item = BlockId>
) -> bool

Handles the main fucntion

The main_id marks the main function.

fn handle_ticking_function(
&mut self,
block_ids: impl Iterator<Item = BlockId>
) -> bool

Handles functions that run every tick

fn handle(&mut self, node: &Node)

fn handle_function(&mut self, function: &Function)

fn handle_fast_store(&mut self, fast_store: &FastStore)

fn handle_fast_store_from_result(
&mut self,
fast_store_from_result: &FastStoreFromResult
)

fn handle_binary_operation(&mut self, binary_operation: &BinaryOperation)

fn handle_condition(&mut self, condition: &Condition)

fn handle_branch(&mut self, branch: &Branch)

fn handle_call(&mut self, call: &Call)

fn handle_execute(&mut self, execute: &ExecuteRaw)

fn handle_write(&mut self, write: &WriteMessage)

fn get_condition(
&mut self,
condition: &Condition,
and_then: Option<MinecraftCommand>
) -> MinecraftCommand

Evaluates this condition and, if it is true, calls and_then. Returns the command instead of adding it to the current stack.

fn get_condition_inner(
&mut self,
condition: &Condition,
parts: &mut Vec<ExecuteComponent>
)

fn get_as_single_command(
&mut self,
commands: Vec<MinecraftCommand>
) -> Option<MinecraftCommand>

Converts a bunch of minecraft commands into a single command

pub fn new(ctx: &'a CompileContext, llir: &'a Llir) -> Self

pub fn build(self) -> Directory